test: move unit tests to inside the main file
