+++ rust/rust-cli/grrs/src/main.rs
@@ -3,7 +3,7 @@ use clap::Parser;
 use log::{info, trace};
 use spinners::{Spinner, Spinners};
 use std::fs::File;
-use std::io::BufReader;
+use std::io::{BufReader, Cursor};
 use std::{thread::sleep, time::Duration};
 
 fn main() -> Result<()> {
@@ -28,4 +28,20 @@ fn main() -> Result<()> {
     portogrrs::find_matches(&mut reader, &args.pattern, &mut std::io::stdout());
 
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn find_a_match() {
+        let input = b"lorem";
+        let mut reader = BufReader::new(Cursor::new(input));
+        let mut writter = Vec::new();
+
+        let result = portogrrs::find_matches(&mut reader, "lorem", &mut writter);
+
+        assert_eq!(result, "lorem");
+    }
 }
